backend/requirements.txt
asn1crypto==0.24.0
cffi==1.11.2
coverage==4.4.2
cryptography==2.1.4
decorator==4.1.2
Django==2.0.1
django-constance==2.0.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-model-utils==3.0.0
django-picklefield==1.0.0
django-x509==0.3.5a0
djangorestframework==3.7.7
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
jsonfield==2.0.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
wcwidth==0.1.7
